Refactor GraphQL mocks
Single function to create all the mocks with separate create functions
for the queries and mutations.
